requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bunnet==1.3.0
Flask==3.0.3
Flask-APScheduler==1.13.1
gunicorn==22.0.0
numpy==1.26.4
pandas==2.2.2
pandas_ta==0.3.14b0
pybit==5.7.0
pymongo==4.7.2
python-dotenv==1.0.1
python-okx==0.2.9
rel==0.4.9.17
requests==2.31.0
websocket-client==1.8.0
